--- a/.github/workflows/trigger-circle-ci-workflow.yml
+++ b/.github/workflows/trigger-circle-ci-workflow.yml
@@ -41,27 +41,27 @@ jobs:
     steps:
       - if: github.event_name == 'pull_request' && ((github.event.action == 'labeled' && github.event.label.name == 'ci:normal') || (github.event.action != 'labeled' && contains(github.event.pull_request.labels.*.name, 'ci:normal')))
         run: |
-          export BRANCH="normal"
-          echo "$BRANCH"
-          echo "branch=$BRANCH" >> $GITHUB_ENV
+          export WORKFLOW="normal"
+          echo "$WORKFLOW"
+          echo "workflow=$WORKFLOW" >> $GITHUB_ENV
       - if: github.event_name == 'pull_request' && ((github.event.action == 'labeled' && github.event.label.name == 'ci:docs') || (github.event.action != 'labeled' && contains(github.event.pull_request.labels.*.name, 'ci:docs')))
         run: |
-          export BRANCH="docs"
-          echo "$BRANCH"
-          echo "branch=$BRANCH" >> $GITHUB_ENV
+          export WORKFLOW="docs"
+          echo "$WORKFLOW"
+          echo "workflow=$WORKFLOW" >> $GITHUB_ENV
       - if: github.event_name == 'push' || (github.event.action == 'labeled' && github.event.label.name == 'ci:merged') || (github.event.action != 'labeled' && contains(github.event.pull_request.labels.*.name, 'ci:merged'))
         run: |
-          export BRANCH="merged"
-          echo "$BRANCH"
-          echo "branch=$BRANCH" >> $GITHUB_ENV
+          export WORKFLOW="merged"
+          echo "$WORKFLOW"
+          echo "workflow=$WORKFLOW" >> $GITHUB_ENV
       - if: github.event_name == 'pull_request' && ((github.event.action == 'labeled' && github.event.label.name == 'ci:daily') || (github.event.action != 'labeled' && contains(github.event.pull_request.labels.*.name, 'ci:daily')))
         run: |
-          export BRANCH="daily"
-          echo "$BRANCH"
-          echo "branch=$BRANCH" >> $GITHUB_ENV
+          export WORKFLOW="daily"
+          echo "$WORKFLOW"
+          echo "workflow=$WORKFLOW" >> $GITHUB_ENV
     outputs:
       workflow: ${{ env.workflow }}
-      ghBaseBranch: ${{ github.event.pull_request.head.ref }}
+      ghBaseBranch: ${{ github.event.pull_request.base.ref }}
       ghPrNumber: ${{ github.event.pull_request.number }}
 
   trigger-circle-ci-workflow: