--- a/scripts/upload-bench.ts
+++ b/scripts/upload-bench.ts
@@ -7,7 +7,7 @@ import { loadBench } from './bench/utils';
 import { SANDBOX_DIRECTORY } from './utils/constants';
 
 const templateKey = process.argv[2];
-let prNumber = process.argv[3];
+const prNumber = process.argv[3];
 
 const GCP_CREDENTIALS = JSON.parse(process.env.GCP_CREDENTIALS || '{}');
 const sandboxDir = process.env.SANDBOX_ROOT || SANDBOX_DIRECTORY;
@@ -87,9 +87,7 @@ const uploadBench = async () => {
   const query = `SELECT * FROM \`storybook-benchmark.benchmark_results.bench2\` WHERE branch='${baseBranch}' AND label='${templateKey}' ORDER BY timestamp DESC LIMIT 1;`;
   const [base]: any[] = await appTable.query(query);
 
-  console.log({ base });
-
-  prNumber = '28508';
+  console.log({ prNumber });
 
   await Promise.all([
     prNumber && prNumber !== '0'