--- a/scripts/upload-bench.ts
+++ b/scripts/upload-bench.ts
@@ -85,11 +85,7 @@ const uploadBench = async () => {
   const baseBranch = 'next';
 
   const query = `SELECT * FROM \`storybook-benchmark.benchmark_results.bench2\` WHERE branch='${baseBranch}' AND label='${templateKey}' ORDER BY timestamp DESC LIMIT 1;`;
-  const [base]: any[] = await appTable.query(query);
-
-  console.log({ base: base, head: row });
-
-  console.log({ prNumber });
+  const [[base]]: any[] = await appTable.query(query);
 
   await Promise.all([
     prNumber && prNumber !== '0'