--- a/scripts/upload-bench.ts
+++ b/scripts/upload-bench.ts
@@ -80,6 +80,10 @@ const uploadBench = async () => {
   const appTable = dataset.table('bench2');
 
   async function uploadToGithub() {
+    if (templateKey !== 'bench/react-vite-default-ts') {
+      console.log('skip uploading results to github because we can only upload 1 result per PR.');
+      return;
+    }
     const [[base]]: any[] = await appTable.query({
       query: `SELECT * FROM \`storybook-benchmark.benchmark_results.bench2\` WHERE branch=@baseBranch AND label=@templateKey ORDER BY timestamp DESC LIMIT 1;`,
       params: { baseBranch, templateKey },